% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{Example Function}
\usage{
example_function(x)
}
\arguments{
\item{x}{A single double value.}
}
\value{
x multiplied by two.
}
\description{
This is an example.
}
